modify presets and test
