+++ cytotable/presets.py
@@ -40,12 +40,12 @@ config = {
                 *
             FROM
                 Image_Filtered AS image
-            LEFT JOIN read_parquet('cytoplasm.parquet') AS cytoplasm ON
+            RIGHT JOIN read_parquet('cytoplasm.parquet') AS cytoplasm ON
                 cytoplasm.Metadata_ImageNumber = image.Metadata_ImageNumber
-            LEFT JOIN read_parquet('cells.parquet') AS cells ON
+            RIGHT JOIN read_parquet('cells.parquet') AS cells ON
                 cells.Metadata_ImageNumber = cytoplasm.Metadata_ImageNumber
                 AND cells.Metadata_ObjectNumber = cytoplasm.Metadata_Cytoplasm_Parent_Cells
-            LEFT JOIN read_parquet('nuclei.parquet') AS nuclei ON
+            RIGHT JOIN read_parquet('nuclei.parquet') AS nuclei ON
                 nuclei.Metadata_ImageNumber = cytoplasm.Metadata_ImageNumber
                 AND nuclei.Metadata_ObjectNumber = cytoplasm.Metadata_Cytoplasm_Parent_Nuclei
             """,
@@ -86,12 +86,12 @@ config = {
                 *
             FROM
                 Per_Image_Filtered AS per_image
-            LEFT JOIN read_parquet('per_cytoplasm.parquet') AS per_cytoplasm ON
+            RIGHT JOIN read_parquet('per_cytoplasm.parquet') AS per_cytoplasm ON
                 per_cytoplasm.Metadata_ImageNumber = per_image.Metadata_ImageNumber
-            LEFT JOIN read_parquet('per_cells.parquet') AS per_cells ON
+            RIGHT JOIN read_parquet('per_cells.parquet') AS per_cells ON
                 per_cells.Metadata_ImageNumber = per_cytoplasm.Metadata_ImageNumber
                 AND per_cells.Cells_Number_Object_Number = per_cytoplasm.Cytoplasm_Parent_Cells
-            LEFT JOIN read_parquet('per_nuclei.parquet') AS per_nuclei ON
+            RIGHT JOIN read_parquet('per_nuclei.parquet') AS per_nuclei ON
                 per_nuclei.Metadata_ImageNumber = per_cytoplasm.Metadata_ImageNumber
                 AND per_nuclei.Nuclei_Number_Object_Number = per_cytoplasm.Cytoplasm_Parent_Nuclei
             """,
@@ -137,12 +137,12 @@ config = {
                 *
             FROM
                 Per_Image_Filtered AS per_image
-            LEFT JOIN read_parquet('per_cytoplasm.parquet') AS per_cytoplasm ON
+            RIGHT JOIN read_parquet('per_cytoplasm.parquet') AS per_cytoplasm ON
                 per_cytoplasm.Metadata_ImageNumber = per_image.Metadata_ImageNumber
-            LEFT JOIN read_parquet('per_cells.parquet') AS per_cells ON
+            RIGHT JOIN read_parquet('per_cells.parquet') AS per_cells ON
                 per_cells.Metadata_ImageNumber = per_cytoplasm.Metadata_ImageNumber
                 AND per_cells.Metadata_Cells_Number_Object_Number = per_cytoplasm.Metadata_Cytoplasm_Parent_Cells
-            LEFT JOIN read_parquet('per_nuclei.parquet') AS per_nuclei ON
+            RIGHT JOIN read_parquet('per_nuclei.parquet') AS per_nuclei ON
                 per_nuclei.Metadata_ImageNumber = per_cytoplasm.Metadata_ImageNumber
                 AND per_nuclei.Metadata_Nuclei_Number_Object_Number = per_cytoplasm.Metadata_Cytoplasm_Parent_Nuclei
             """,
@@ -191,14 +191,14 @@ config = {
                 *
             FROM
                 Image_Filtered AS image
-            LEFT JOIN read_parquet('cytoplasm.parquet') AS cytoplasm ON
+            RIGHT JOIN read_parquet('cytoplasm.parquet') AS cytoplasm ON
                 cytoplasm.Metadata_TableNumber = image.Metadata_TableNumber
                 AND cytoplasm.Metadata_ImageNumber = image.Metadata_ImageNumber
-            LEFT JOIN read_parquet('cells.parquet') AS cells ON
+            RIGHT JOIN read_parquet('cells.parquet') AS cells ON
                 cells.Metadata_TableNumber = cytoplasm.Metadata_TableNumber
                 AND cells.Metadata_ImageNumber = cytoplasm.Metadata_ImageNumber
                 AND cells.Cells_ObjectNumber = cytoplasm.Metadata_Cytoplasm_Parent_Cells
-            LEFT JOIN read_parquet('nuclei.parquet') AS nuclei ON
+            RIGHT JOIN read_parquet('nuclei.parquet') AS nuclei ON
                 nuclei.Metadata_TableNumber = cytoplasm.Metadata_TableNumber
                 AND nuclei.Metadata_ImageNumber = cytoplasm.Metadata_ImageNumber
                 AND nuclei.Nuclei_ObjectNumber = cytoplasm.Metadata_Cytoplasm_Parent_Nuclei
