+++ tests/test_convert.py
@@ -1115,25 +1115,3 @@ def test_in_carta_to_parquet(
         # check the data against one another
         assert cytotable_result_table.schema.equals(ddb_result.schema)
         assert cytotable_result_table.shape == ddb_result.shape
-
-
-def test_avoid_na_row_output(
-    load_parsl_default: None, fx_tempdir: str, data_dir_cellprofiler: str
-):
-    """
-    Test to help detect and avoid scenarios where rows of NA-based data are returned
-    due to undetected objects in compartments (and as a result, imagenumbers).
-    For example, if there are imagenumbers in the image table and not the compartment table,
-    we want to avoid returning rows of NA data from the compartment tables after joins take place. 
-    """
-
-    # run convert using a dataset known to contain the scenario outlined above.
-    parquet_file = convert(
-        source_path=f"{data_dir_cellprofiler}/nf1_cellpainting_data/test-Plate_3_nf1_analysis.sqlite",
-        dest_path=f"{fx_tempdir}/nf1_cellpainting_data/test-Plate_3_nf1_analysis.parquet",
-        dest_datatype="parquet",
-        preset="cellprofiler_sqlite_pycytominer",
-    )
-
-    # check for a column containing any null data
-    assert not pc.is_null(parquet.read_table(parquet_file).column("Metadata_ImageNumber"))
